Refactoring

+++ src/main/java/com/logines/schedule/controller/UserController.java
@@ -38,7 +38,7 @@ public class UserController {
     }
 
     @PostMapping("/registration")
-    public String registration(@PathVariable("id") long id,
+    public String registration(@PathVariable("id") Integer id,
                                @Valid Users users,
                                BindingResult bindingResult,
                                Model model) {
